adjusted the redirect test as per new changes and comments from Ripal
